requirements.txt
pytest==8.0.0
selenium==4.18.1
webdriver-manager==4.0.1
pytest-html==4.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.1
requests==2.31.0
allure-pytest==2.13.2
PyYAML==6.0.1
pytest-rerunfailures==12.0
pytest-sugar==0.9.7 